+++ RenderWidget.h
@@ -22,9 +22,9 @@ public:
 
   void myDrawLine(float x1, float y1, float x2, float y2);
   void drawTicks(float xc, float yc, float r);
-  void mapPoint(int &x, int &y);
-  void drawLine(int x1, int y1, int x2, int y2);
-  void drawLineFromDerivation(int x1, int y1, int x2, int y2);
+  void mapPoint(float &x, float &y);
+  void drawLine(float x1, float y1, float x2, float y2);
+  void drawLineFromDerivation(float x1, float y1, float x2, float y2);
   void drawCircle(float xc, float yc, float r);
   void drawArc(float xc, float yc, float t1, float t2, float r);
   void drawEllipse(float xc, float yc, float a, float b);
